+++ onnxruntime/core/providers/cuda/cuda_allocator.h
@@ -3,7 +3,8 @@
 
 #pragma once
 
-#include <unordered_set>
+
+#include "core/common/inlined_containers.h"
 #include "core/framework/allocator.h"
 #include "core/platform/ort_mutex.h"
 
@@ -47,7 +48,7 @@ class CUDAExternalAllocator : public CUDAAllocator {
   ExternalAlloc alloc_;
   ExternalFree free_;
   ExternalEmptyCache empty_cache_;
-  std::unordered_set<void*> reserved_;
+  InlinedHashSet<void*> reserved_;
 };
 
 //TODO: add a default constructor
