+++ orttraining/orttraining/training_ops/cuda/math/isfinite.cc
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 #include "orttraining/training_ops/cuda/math/isfinite.h"
+#include "orttraining/training_ops/cuda/math/isfinite_impl.h"
 
 using namespace ONNX_NAMESPACE;
 using namespace onnxruntime::common;
