+++ orttraining/orttraining/training_ops/cuda/math/isfinite_impl.cu
@@ -4,6 +4,7 @@
 #include <cuda_fp16.h>
 #include "core/providers/cuda/cu_inc/common.cuh"
 #include "contrib_ops/cuda/math/isfinite.cuh"
+#include "isfinite_impl.h"
 
 namespace onnxruntime {
 namespace cuda {
