+++ orttraining/orttraining/training_ops/cuda/math/mixed_precision_scale.h
@@ -8,14 +8,6 @@
 namespace onnxruntime {
 namespace cuda {
 
-template <typename SrcT, typename DstT>
-void Impl_MixedPrecisionScale(
-    cudaStream_t stream,
-    const SrcT* input_data,
-    const float* scale_data,
-    DstT* output_data,
-    size_t count);
-
 template <typename SrcT>
 class MixedPrecisionScale final : public CudaKernel {
  public:
