+++ orttraining/orttraining/training_ops/cuda/math/scale.h
@@ -7,23 +7,6 @@
 namespace onnxruntime {
 namespace cuda {
 
-template <typename ScaleT>
-struct GetScaleValueImpl {
-  void operator()(const Tensor* scale, float& scale_value) const {
-    ORT_ENFORCE(scale->Shape().Size() == 1, "Scale input should have a single value.");
-    scale_value = static_cast<float>(*(scale->template Data<ScaleT>()));
-    ORT_ENFORCE(scale_value != 0.0f, "Scale value must not be 0.");
-  }
-};
-
-template <typename T>
-void Impl_Scale(
-    cudaStream_t stream,
-    const T* input_data,
-    const float scale_value,
-    T* output_data,
-    size_t count);
-
 template <typename T>
 class Scale final : public CudaKernel {
  public:
