+++ orttraining/orttraining/training_ops/cuda/math/softmax_grad.h
@@ -8,9 +8,6 @@
 namespace onnxruntime {
 namespace cuda {
 
-template <typename input_t, typename output_t, typename acc_t, bool is_log_softmax>
-void dispatch_softmax_backward(cudaStream_t stream, output_t* grad_input, const input_t* grad, const input_t* output, int softmax_elements, int softmax_elements_stride, int batch_count);
-
 template <typename T>
 class SoftmaxGrad final : public CudaKernel {
  public:
