+++ orttraining/orttraining/training_ops/cuda/optimizer/adam.cc
@@ -6,6 +6,7 @@
 #include "core/providers/cuda/math/binary_elementwise_ops.h"
 #include "orttraining/training_ops/cuda/optimizer/common.h"
 #include "orttraining/training_ops/cuda/optimizer/adam.h"
+#include "orttraining/training_ops/cuda/optimizer/adam_impl.h"
 
 namespace onnxruntime {
 namespace cuda {
