+++ orttraining/orttraining/training_ops/cuda/optimizer/gradient_control.h
@@ -22,14 +22,5 @@ class InPlaceAccumulator final : public CudaKernel {
   Status ComputeInternal(OpKernelContext* context) const override;
 };
 
-// Implementation can be found in cuda file, optimizers_impl.cu
-template <typename T, typename T_GRAD>
-void InPlaceAccumulatorImpl(
-    cudaStream_t stream,
-    const T* gradient_buffer,
-    const T_GRAD* gradient,
-    T* accumulated_gradient,
-    size_t count);
-
 }  // namespace cuda
 }  // namespace onnxruntime
