+++ orttraining/orttraining/training_ops/cuda/optimizer/lamb.cc
@@ -7,6 +7,7 @@
 #include "core/providers/cuda/math/binary_elementwise_ops.h"
 #include "orttraining/training_ops/cuda/optimizer/common.h"
 #include "orttraining/training_ops/cuda/optimizer/lamb.h"
+#include "orttraining/training_ops/cuda/optimizer/lamb_impl.h"
 
 namespace onnxruntime {
 namespace cuda {
