+++ orttraining/orttraining/training_ops/cuda/optimizer/sg.h
@@ -8,16 +8,6 @@
 namespace onnxruntime {
 namespace cuda {
 
-template <typename T>
-void SGDOptimizerImpl(
-    cudaStream_t stream,
-    const T* eta,
-    const T* weights,
-    const T* gradients,
-    T* weight_out,
-    T* gradients_out,
-    size_t count);
-
 class SGDOptimizer final : public CudaKernel {
  public:
   SGDOptimizer(const OpKernelInfo& info) : CudaKernel(info) {}
