+++ orttraining/orttraining/training_ops/cuda/reduction/all_impl.cu
@@ -1,11 +1,12 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include "orttraining/training_ops/cuda/reduction/all.h"
-
 #include <thrust/logical.h>
 #include <thrust/functional.h>
 #include <thrust/execution_policy.h>
+
+#include "orttraining/training_ops/cuda/reduction/all_impl.h"
+
 #ifdef _WIN32
 #pragma warning(disable : 4244)
 #endif
